fix data type for rx and tx in set_buffer_size for tty

+++ src/posix/tty.rs
@@ -675,7 +675,7 @@ impl SerialPort for TTYPort {
         Ok(())
     }
 
-    fn set_buffer_size(&mut self, _rx: u32, _tx: u32) -> Result<()> {
+    fn set_buffer_size(&mut self, _rx: usize, _tx: usize) -> Result<()> {
         Err(Error::new(
             ErrorKind::Unsupported,
             "Configurable buffer sizes are not supported on this platform",
